% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_prediction}
\alias{get_prediction}
\title{Retuns a preduction based on a provided Weight}
\usage{
get_prediction(x)
}
\arguments{
\item{x}{Value for the weight of the car}
}
\description{
Retuns a preduction based on a provided Weight
}
